refactor: PascalCase 파일명 49개를 kebab-case로 통일
컴포넌트 파일 네이밍 컨벤션을 Next.js/FSD 표준인 kebab-case로 통일:
- analytics/ 11개, counseling/ 22개, statistics/ 10개
- matching/ 4개, satisfaction/ 1개, teachers/ 1개
- 약어는 단일 단어로 처리 (MBTI→mbti, AI→ai, CSV→csv)
- git mv로 히스토리 보존, 모든 import 경로 동기화
