+++ src/app/[locale]/(dashboard)/matching/matching-tabs.tsx
@@ -14,7 +14,7 @@ import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs'
 import { ManualAssignmentForm } from '@/components/assignment/manual-assignment-form'
 import { BatchAssignment } from '@/components/assignment/batch-assignment'
 import { TeacherAssignmentTable } from '@/components/assignment/teacher-assignment-table'
-import { MatchingHistoryTab } from '@/components/matching/MatchingHistoryTab'
+import { MatchingHistoryTab } from '@/components/matching/matching-history-tab'
 import { SmartMatchingSection } from '@/components/matching/smart-matching-section'
 import type { UnassignedStudent, AssignedStudent } from '@/components/matching/unassigned-student-combobox'
 import type { LlmProviderOption } from '@/components/matching/matching-model-selector'
