+++ src/app/[locale]/(dashboard)/teachers/[id]/students/page.tsx
@@ -2,7 +2,7 @@ import { notFound } from 'next/navigation'
 import { verifySession } from '@/lib/dal'
 import { getRBACPrisma } from '@/lib/db/common/rbac'
 import { getTeacherStudents, getTeacherStudentMetrics } from '@/lib/actions/teacher/performance'
-import { TeacherStudentList } from '@/components/teachers/TeacherStudentList'
+import { TeacherStudentList } from '@/components/teachers/teacher-student-list'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import Link from 'next/link'
