+++ src/app/[locale]/(dashboard)/teams/[id]/composition/page.tsx
@@ -1,7 +1,7 @@
 import { notFound } from "next/navigation"
 import { analyzeTeamComposition } from "@/lib/analysis/team-composition"
 import { getTeamById } from "@/lib/actions/common/teams"
-import { TeamCompositionPanel } from "@/components/analytics/TeamCompositionPanel"
+import { TeamCompositionPanel } from "@/components/analytics/team-composition-panel"
 import { Card, CardContent } from "@/components/ui/card"
 
 interface PageProps {
