+++ src/components/analytics/control-variable-panel.tsx
