+++ src/components/analytics/mbti-distribution-chart.tsx
