+++ src/components/analytics/performance-dashboard.tsx
@@ -5,11 +5,11 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Button } from "@/components/ui/button"
-import { PerformanceMetricsGrid } from "./PerformanceMetricsGrid"
-import { GradeTrendChart, TrendDataPoint } from "./GradeTrendChart"
-import { MultiSubjectChart, SubjectAverage } from "./MultiSubjectChart"
-import { ControlVariablePanel } from "./ControlVariablePanel"
-import { TeacherWithMetrics } from "./TeacherPerformanceCard"
+import { PerformanceMetricsGrid } from "./performance-metrics-grid"
+import { GradeTrendChart, TrendDataPoint } from "./grade-trend-chart"
+import { MultiSubjectChart, SubjectAverage } from "./multi-subject-chart"
+import { ControlVariablePanel } from "./control-variable-panel"
+import { TeacherWithMetrics } from "./teacher-performance-card"
 import { BarChart3, Users, TrendingUp, FileDown } from "lucide-react"
 import { CounselingStats } from "@/lib/actions/common/analytics"
