+++ src/components/counseling/ai-summary-generator.tsx
