+++ src/components/counseling/compatibility-score-card.tsx
