+++ src/components/counseling/counseling-page-tabs.tsx
@@ -5,9 +5,9 @@ import { useRouter } from "next/navigation"
 import { Button } from "@/components/ui/button"
 import { Plus } from "lucide-react"
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs"
-import { ReservationList } from "@/components/counseling/ReservationList"
-import { ReservationForm } from "@/components/counseling/ReservationForm"
-import { ReservationCalendarView } from "@/components/counseling/ReservationCalendarView"
+import { ReservationList } from "@/components/counseling/reservation-list"
+import { ReservationForm } from "@/components/counseling/reservation-form"
+import { ReservationCalendarView } from "@/components/counseling/reservation-calendar-view"
 import { getReservationsAction } from "@/lib/actions/counseling/reservations-query"
 import type { ReservationWithRelations } from "@/types/counseling"
 import type { CounselingSessionData } from "./types"
