+++ src/components/counseling/counseling-session-form.tsx
@@ -18,7 +18,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { recordCounselingAction } from "@/lib/actions/common/performance"
 import { counselingSchema, type CounselingFormData } from "@/lib/validations/counseling"
 import { toast } from "sonner"
-import { AISupportPanel } from "./AISupportPanel"
+import { AISupportPanel } from "./ai-support-panel"
 
 interface CounselingSessionFormProps {
   studentId: string
