+++ src/components/counseling/personality-summary-card.tsx
