+++ src/components/matching/matching-history-tab.tsx
@@ -6,8 +6,8 @@ import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
-import { MatchingAuditTable } from '@/components/matching/MatchingAuditTable'
-import { AuditLogDetailDialog } from '@/components/matching/AuditLogDetailDialog'
+import { MatchingAuditTable } from '@/components/matching/matching-audit-table'
+import { AuditLogDetailDialog } from '@/components/matching/audit-log-detail-dialog'
 import { getMatchingHistory, type MatchingHistoryParams } from '@/lib/actions/matching/history'
 import { getTeachers } from '@/lib/actions/teacher/crud'
 import type { AuditLogEntry } from '@/lib/actions/admin/audit'
