+++ src/components/statistics/counseling-trend-chart.tsx
