+++ src/components/statistics/counseling-type-chart.tsx
